test(rl): move shared process helper into the unit_test lane
tests/_util/ failed the test-layout gate (Python test files must live under
tests/unit_test/ or tests/test_model/). Move the stdlib-only subprocess line
helper to tests/unit_test/_util/process.py (kept under the lightweight lane so
the unit test still imports it without heavy deps) and repoint both consumers.
