+++ tests/unit_test/test_process_line_wait.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import subprocess
 import sys
 
-from tests._util.process import _wait_for_process_line
+from tests.unit_test._util.process import _wait_for_process_line
 
 
 def test_wait_for_process_line_keeps_prefetched_stdout_lines():
